fix: context editor — system prompt override, mid-session support, UI improvements
- Fix system prompt override not taking effect: the SDK's AgentSession
  rebuilds _baseSystemPrompt on every prompt() call, overwriting direct
  state.systemPrompt mutations. Now overwrite _baseSystemPrompt on the
  session instance so overrides survive across turns.

- Move context editor trigger from header icon into the '+' action menu
  alongside 'Add photos or files', making it discoverable at all times.

- Remove new-session-only restriction — context can now be edited
  mid-session (only cost is a one-time cache miss on the next turn).

- Reset editor state to defaults on open so stale overrides from a
  previous session don't bleed into new sessions.

- Fix layout overflow: extract ToolRow/SkillRow/ContextSection into
  context-editor-parts.tsx, add proper min-w-0/overflow constraints,
  wrap long descriptions instead of truncating.

- Widen dialog (58rem), taller textarea (h-80), larger scroll area (65vh).
